move sensorfilter functions into files
so its not a giant file, and consolidate usings/includes. some other small optimisations
