+++ module/input/SensorFilter/src/sensorfilter/update_odometry.cpp
@@ -1,0 +1,114 @@
+#include "SensorFilter.hpp"
+
+#include "utility/input/FrameID.hpp"
+#include "utility/math/euler.hpp"
+
+namespace module::input {
+
+    using message::behaviour::state::Stability;
+    using message::behaviour::state::WalkState;
+    using message::input::Sensors;
+    using message::platform::RawSensors;
+
+    using utility::input::FrameID;
+    using utility::math::euler::mat_to_rpy_intrinsic;
+    using utility::math::euler::rpy_intrinsic_to_mat;
+
+    void SensorFilter::update_odometry(std::unique_ptr<Sensors>& sensors,
+                                       const std::shared_ptr<const Sensors>& previous_sensors,
+                                       const RawSensors& raw_sensors,
+                                       const Stability& stability) {
+        // Use ground truth instead of calculating odometry, then return
+        if (cfg.use_ground_truth) {
+            // Construct world {w} to torso {t} space transform from ground truth.inverse();
+            sensors->Htw = Eigen::Isometry3d(raw_sensors.odometry_ground_truth.Htw);
+            // Construct robot {r} to world {w} space transform from ground truth
+            Eigen::Isometry3d Hrw = Eigen::Isometry3d::Identity();
+            Hrw.linear() = Eigen::AngleAxisd(mat_to_rpy_intrinsic(sensors->Htw.linear()).z(), Eigen::Vector3d::UnitZ())
+                               .toRotationMatrix();
+            Hrw.translation() = Eigen::Vector3d(sensors->Htw.translation().x(), sensors->Htw.translation().y(), 0.0);
+            sensors->Hrw      = Hrw;
+            sensors->vTw      = raw_sensors.odometry_ground_truth.vTw;
+            return;
+        }
+
+        // Compute time since last update
+        const double dt = std::max(
+            std::chrono::duration_cast<std::chrono::duration<double>>(
+                raw_sensors.timestamp - (previous_sensors ? previous_sensors->timestamp : raw_sensors.timestamp))
+                .count(),
+            0.0);
+
+        // Perform Mahony update
+        const auto Rwt_mahony      = mahony_filter.update(sensors->accelerometer, sensors->gyroscope, dt);
+        Eigen::Vector3d rpy_mahony = mat_to_rpy_intrinsic(Rwt_mahony);
+
+        // If fallen, keep position still
+        if (stability <= Stability::FALLING) {
+            Eigen::Isometry3d Hwt =
+                previous_sensors == nullptr ? Eigen::Isometry3d::Identity() : previous_sensors->Htw.inverse();
+            // Htw rotation is combination of Mahony pitch and roll and existing yaw
+            Hwt.linear() = Rwt_mahony;
+            sensors->Htw = Hwt.inverse();
+
+            // Get robot to world
+            sensors->Hrw = previous_sensors == nullptr ? Eigen::Isometry3d::Identity() : previous_sensors->Hrw;
+
+            // Set velocity to zero
+            sensors->vTw = Eigen::Vector3d::Zero();
+
+            return;
+        }
+
+        // If sensors detected a new foot phase, update the anchor frame
+        if (planted_anchor_foot != sensors->planted_foot_phase
+            && sensors->planted_foot_phase != WalkState::Phase::DOUBLE) {
+            switch (planted_anchor_foot.value) {
+                case WalkState::Phase::RIGHT:
+                    Hwp = Hwp * sensors->Htx[FrameID::R_FOOT_BASE].inverse() * sensors->Htx[FrameID::L_FOOT_BASE];
+                    break;
+                case WalkState::Phase::LEFT:
+                    Hwp = Hwp * sensors->Htx[FrameID::L_FOOT_BASE].inverse() * sensors->Htx[FrameID::R_FOOT_BASE];
+                    break;
+                default: log<WARN>("Anchor frame should not be updated in double support phase"); break;
+            }
+            planted_anchor_foot = sensors->planted_foot_phase;
+            // Set the z translation, roll and pitch of the anchor frame to 0 as known to be on field plane
+            Hwp.translation().z() = 0;
+            Hwp.linear()          = rpy_intrinsic_to_mat(Eigen::Vector3d(0, 0, mat_to_rpy_intrinsic(Hwp.linear()).z()));
+        }
+        sensors->Hwp = Hwp;
+
+        // Compute torso pose using kinematics from anchor frame (current planted foot)
+        const Eigen::Isometry3d Hpt        = planted_anchor_foot.value == WalkState::Phase::RIGHT
+                                                 ? Eigen::Isometry3d(sensors->Htx[FrameID::R_FOOT_BASE].inverse())
+                                                 : Eigen::Isometry3d(sensors->Htx[FrameID::L_FOOT_BASE].inverse());
+        const Eigen::Isometry3d Hwt_anchor = Hwp * Hpt;
+
+        // Construct world {w} to torso {t} space transform (mahony orientation, anchor translation)
+        Eigen::Isometry3d Hwt = Eigen::Isometry3d::Identity();
+        Hwt.translation()     = Hwt_anchor.translation();
+        Hwt.linear()          = Rwt_mahony;
+        sensors->Htw          = Hwt.inverse();
+
+        // Construct robot {r} to world {w} space transform (just x-y translation and yaw rotation)
+        Eigen::Isometry3d Hwr = Eigen::Isometry3d::Identity();
+        Hwr.linear()          = Eigen::AngleAxisd(rpy_mahony.z(), Eigen::Vector3d::UnitZ()).toRotationMatrix();
+        Hwr.translation()     = Eigen::Vector3d(Hwt_anchor.translation().x(), Hwt_anchor.translation().y(), 0.0);
+        sensors->Hrw          = Hwr.inverse();
+
+        // Low pass filter for torso velocity
+        const double y_current     = Hwt.translation().y();
+        const double y_prev        = previous_sensors ? previous_sensors->Htw.inverse().translation().y() : y_current;
+        const double y_dot_current = (y_current - y_prev) / dt;
+        const double y_dot =
+            (dt / cfg.y_cut_off_frequency) * y_dot_current + (1 - (dt / cfg.y_cut_off_frequency)) * sensors->vTw.y();
+        const double x_current     = Hwt.translation().x();
+        const double x_prev        = previous_sensors ? previous_sensors->Htw.inverse().translation().x() : x_current;
+        const double x_dot_current = (x_current - x_prev) / dt;
+        const double x_dot =
+            (dt / cfg.x_cut_off_frequency) * x_dot_current + (1 - (dt / cfg.x_cut_off_frequency)) * sensors->vTw.x();
+        sensors->vTw = Eigen::Vector3d(x_dot, y_dot, 0);
+    }
+
+}  // namespace module::input
